requirements.txt
blinker==1.8.2
cachelib==0.9.0
click==8.1.7
Deprecated==1.2.15
distlib==0.3.9
ecdsa==0.19.0
filelock==3.16.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-Limiter==3.10.1
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
marshmallow-sqlalchemy==1.3.0
mdurl==0.1.2
mysql-connector-python==8.4.0
ordered-set==4.1.0
packaging==24.1
platformdirs==4.3.6
pyasn1==0.6.1
Pygments==2.19.1
python-jose==3.3.0
PyYAML==6.0.2
rich==13.9.4
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
virtualenv==20.26.6
Werkzeug==3.0.4
wrapt==1.17.2
